convert to big decimal

+++ java/src/main/java/com/techelevator/application/Candy.java
@@ -1,13 +1,15 @@
 package com.techelevator.application;
 
+import java.math.BigDecimal;
+
 public class Candy extends VendingItems {
 
-    public Candy(String slotNo, String nameOfProduct, int price) {
+    public Candy(String slotNo, String nameOfProduct, BigDecimal price) {
         super(slotNo, nameOfProduct, price);
     }
 
     @Override
-    public String dispenseMessage(boolean isDiscount, int remainingBalance) {
+    public String dispenseMessage(boolean isDiscount, BigDecimal remainingBalance) {
         System.out.println(getNameOfProduct() + " " + getPrice(isDiscount) + " " + remainingBalance);
         System.out.println("Sugar, Sugar, so Sweet!");
         return "Sugar, Sugar, so Sweet!";
